organized components into sub folders
