+++ src/Components/App/HomePage.js
@@ -3,6 +3,7 @@
 function HomePage() {
     return (
         <div>
+            <h1>Welcome to inIT!</h1>
             Home Page
         </div>
     );
